requirements.txt
numpy<2.0.0
pandas
scipy
scikit-learn
scikit-surprise
joblib
